pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-redis>=0.2.1",
    "a2a-sdk>=0.3.24",
    "fastapi[standard]>=0.134.0",
    "redis>=7.3.0",
    "sse-starlette>=3.3.2",
    "uvicorn>=0.41.0",
]

[dependency-groups]
dev = [
    "anyio>=4.12.1",
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
